Move some logic that was in the template into code that can be easily unit tested.
