+++ lib/meetupinator.rb
@@ -4,6 +4,7 @@ require 'meetupinator/event_finder'
 require 'meetupinator/event_list_file_reader'
 require 'meetupinator/event_list_file_writer'
 require 'meetupinator/formatter'
-require 'meetupinator/meetup_api'
-require 'meetupinator/version'
 require 'meetupinator/input_file_reader'
+require 'meetupinator/meetup_api'
+require 'meetupinator/template_parameters'
+require 'meetupinator/version'
